+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 1.0.2
+
+Fixes for "sources" command.
+Added: "version" subcommand, "-v" and "--version" options.
+
 # 1.0.1
 
 Fixes. Update README.